Move CLI runtime selection out of lib
